Add files via upload
This project is a web-based application designed to detect skin diseases using deep learning. It utilizes a Convolutional Neural Network (CNN) model, built and trained with TensorFlow, to classify images into one of four types of skin diseases: acne, impetigo, melasma, and rosacea. The application provides a user-friendly interface developed with Flask, allowing users to upload an image of the affected skin area. The system processes the input image and predicts the type of disease, displaying the result in an intuitive manner. This project demonstrates the integration of machine learning with web technologies and aims to assist in preliminary diagnosis, potentially supporting dermatologists and medical practitioners in their work. It leverages Python, TensorFlow, and Flask to deliver an efficient and accurate solution for skin disease detection.